Convert and Save sequence data frame to fasta file

Convert and Save sequence data frame to fasta file.

dat2fasta(dat, outfile = "out.fasta")

Arguments

dat
data frame by read.phylip or read.fasta

outfile
a character string, representing the name of the fasta file to be generated

Details

The column of the data frame must be: 1. seq.name, 2. seq.text, represent the name of the sequences, the content of the sequence, eg. ATCGGGAAC.

Value

This is a routine without return value.

References

http://www.genomatix.de/online_help/help/sequence_formats.html

Author

Jinlong Zhang <jinlongzhang01@gmail.com>

See also

Examples

cat(
">seq_2", "GTCTTATAAGAAAGAATAAGAAAG--AAATACAAA-------AAAAAAGA",
">seq_3", "GTCTTATAAGAAAGAAATAGAAAAGTAAAAAAAAA-------AAAAAAAG",
">seq_5", "GACATAAGACATAAAATAGAATACTCAATCAGAAACCAACCCATAAAAAC",
">seq_8", "ATTCCAAAATAAAATACAAAAAGAAAAAACTAGAAAGTTTTTTTTCTTTG",
">seq_9", "ATTCTTTGTTCTTTTTTTTCTTTAATCTTTAAATAAACCTTTTTTTTTTA",
file = "trn1.fasta", sep = "\n")

res <- read.fasta("trn1.fasta")
dat2fasta(res)

#> out.fasta has been saved to  /Users/jinlong/Documents/github/phylotools/docs/reference 

unlink("trn1.fasta")
unlink("out.fasta")