get the names of sequences from a PHYLIP file

get the names of sequences from a PHYLIP file.

get.phylip.name(infile, clean_name = FALSE)

Arguments

infile
character representing the name or path of the phylip file.

clean_name
logical, representing cleaning of the names will be performed.

Details

Punctuation characters and white space be replaced by "_". Definition of Punctuation characters can be found at regex.

Value

a character vector of the names of the sequences

Author

Jinlong Zhang <jinlongzhang01@gmail.com>

See also

Examples

  cat("6 22",
  "seq_1    --TTACAAATTGACTTATTATA",
  "seq_2    GATTACAAATTGACTTATTATA",
  "seq_3    GATTACAAATTGACTTATTATA",
  "seq_5    GATTACAAATTGACTTATTATA",
  "seq_8    GATTACAAATTGACTTATTATA",
  "seq_10   ---TACAAATTGAATTATTATA",
  file = "matk.phy", sep = "\n")
get.phylip.name("matk.phy")

#> [1] "seq_1"  "seq_2"  "seq_3"  "seq_5"  "seq_8"  "seq_10"

unlink("matk.phy")